Change AbstractOutputPreset to AbstractOutput
